requirements.txt
aiogram==3.14.0
betterlogging==1.0.0
orjson==3.10.11
python-dotenv==1.0.1aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
notion-client==2.2.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
yarl==1.18.0
